Trait pallet_contracts::api_doc::Version2

pub trait Version2 {
    fn set_storage(
        key_ptr: u32,
        key_len: u32,
        value_ptr: u32,
        value_len: u32
    ) -> Result<u32, TrapReason>;
    fn seal_set_storage(
        key_ptr: u32,
        key_len: u32,
        value_ptr: u32,
        value_len: u32
    ) -> Result<u32, TrapReason>;
}

All functions available in the seal2 module

Required Methods

fn set_storage(
key_ptr: u32,
key_len: u32,
value_ptr: u32,
value_len: u32
) -> Result<u32, TrapReason>

Set the value at the given key in the contract storage.

The key and value lengths must not exceed the maximums defined by the contracts module parameters. Specifying a value_len of zero will store an empty value.

Parameters

key_ptr: pointer into the linear memory where the location to store the value is placed.

key_len: the length of the key in bytes.

value_ptr: pointer into the linear memory where the value to set is placed.

value_len: the length of the value in bytes.

Return Value

Returns the size of the pre-existing value at the specified key if any. Otherwise SENTINEL is returned as a sentinel value.

Wasm Import Statement

(import "seal2" "set_storage" (func ...))

fn seal_set_storage(
key_ptr: u32,
key_len: u32,
value_ptr: u32,
value_len: u32
) -> Result<u32, TrapReason>

This is just an alias function to set_storage() with backwards-compatible prefixed identifier.

Wasm Import Statement

(import "seal2" "seal_set_storage" (func ...))