Trait pallet_contracts::chain_extension::Ext

pub trait Ext: Sealed {
    type T: Config;

    fn call(
        &mut self, 
        gas_limit: Weight, 
        to: <Self::T as Config>::AccountId, 
        value: <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance, 
        input_data: Vec<u8>, 
        allows_reentry: bool
    ) -> Result<ExecReturnValue, ExecError>;

    fn delegate_call(
        &mut self, 
        code: <Self::T as Config>::Hash, 
        input_data: Vec<u8>
    ) -> Result<ExecReturnValue, ExecError>;

    fn instantiate(
        &mut self, 
        gas_limit: Weight, 
        code: <Self::T as Config>::Hash, 
        value: <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance, 
        input_data: Vec<u8>, 
        salt: &[u8]
    ) -> Result<(<Self::T as Config>::AccountId, ExecReturnValue), ExecError>;

    fn terminate(
        &mut self, 
        beneficiary: &<Self::T as Config>::AccountId
    ) -> Result<(), DispatchError>;

    fn transfer(
        &mut self, 
        to: &<Self::T as Config>::AccountId, 
        value: <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance
    ) -> DispatchResult;

    fn get_storage(&mut self, key: &[u8; 32]) -> Option<Vec<u8>>;

    fn get_storage_size(&mut self, key: &[u8; 32]) -> Option<u32>;

    fn set_storage(
        &mut self, 
        key: [u8; 32], 
        value: Option<Vec<u8>>, 
        take_old: bool
    ) -> Result<WriteOutcome, DispatchError>;

    fn caller(&self) -> &<Self::T as Config>::AccountId;

    fn is_contract(&self, address: &<Self::T as Config>::AccountId) -> bool;

    fn caller_is_origin(&self) -> bool;

    fn address(&self) -> &<Self::T as Config>::AccountId;

    fn balance(
        &self
    ) -> <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance;

    fn value_transferred(
        &self
    ) -> <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance;

    fn now(&self) -> &<<Self::T as Config>::Time as Time>::Moment;

    fn minimum_balance(
        &self
    ) -> <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance;

    fn random(
        &self, 
        subject: &[u8]
    ) -> (<Self::T as Config>::Hash, <Self::T as Config>::BlockNumber);

    fn deposit_event(
        &mut self, 
        topics: Vec<<Self::T as Config>::Hash>, 
        data: Vec<u8>
    );

    fn block_number(&self) -> <Self::T as Config>::BlockNumber;

    fn max_value_size(&self) -> u32;

    fn get_weight_price(
        &self, 
        weight: Weight
    ) -> <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance;

    fn schedule(&self) -> &Schedule<Self::T>;

    fn gas_meter(&mut self) -> &mut GasMeter<Self::T>;

    fn append_debug_buffer(&mut self, msg: &str) -> bool;

    fn call_runtime(
        &self, 
        call: <Self::T as Config>::Call
    ) -> DispatchResultWithPostInfo;

    fn ecdsa_recover(
        &self, 
        signature: &[u8; 65], 
        message_hash: &[u8; 32]
    ) -> Result<[u8; 33], ()>;

}

An interface that provides access to the external environment in which the smart-contract is executed.

This interface is specialized to an account of the executing code, so all operations are implicitly performed on that account.

Note

This trait is sealed and cannot be implemented by downstream crates.

Associated Types

type T: Config

Required methods

fn call(
&mut self,
gas_limit: Weight,
to: <Self::T as Config>::AccountId,
value: <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance,
input_data: Vec<u8>,
allows_reentry: bool
) -> Result<ExecReturnValue, ExecError>

Call (possibly transferring some amount of funds) into the specified account.

Returns the original code size of the called contract.

Return Value

Result<(ExecReturnValue, CodeSize), (ExecError, CodeSize)>

fn delegate_call(
&mut self,
code: <Self::T as Config>::Hash,
input_data: Vec<u8>
) -> Result<ExecReturnValue, ExecError>

Execute code in the current frame.

Returns the original code size of the called contract.

Return Value

Result<ExecReturnValue, ExecError>

fn instantiate(
&mut self,
gas_limit: Weight,
code: <Self::T as Config>::Hash,
value: <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance,
input_data: Vec<u8>,
salt: &[u8]
) -> Result<(<Self::T as Config>::AccountId, ExecReturnValue), ExecError>

Instantiate a contract from the given code.

Returns the original code size of the called contract. The newly created account will be associated with code. value specifies the amount of value transferred from this to the newly created account.

Return Value

Result<(AccountId, ExecReturnValue, CodeSize), (ExecError, CodeSize)>

fn terminate(
&mut self,
beneficiary: &<Self::T as Config>::AccountId
) -> Result<(), DispatchError>

Transfer all funds to beneficiary and delete the contract.

Since this function removes the self contract eagerly, if succeeded, no further actions should be performed on this Ext instance.

This function will fail if the same contract is present on the contract call stack.

fn transfer(
&mut self,
to: &<Self::T as Config>::AccountId,
value: <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance
) -> DispatchResult

Transfer some amount of funds into the specified account.

fn get_storage(&mut self, key: &[u8; 32]) -> Option<Vec<u8>>

Returns the storage entry of the executing account by the given key.

Returns None if the key wasn’t previously set by set_storage or was deleted.

fn get_storage_size(&mut self, key: &[u8; 32]) -> Option<u32>

Returns Some(len) (in bytes) if a storage item exists at key.

Returns None if the key wasn’t previously set by set_storage or was deleted.

fn set_storage(
&mut self,
key: [u8; 32],
value: Option<Vec<u8>>,
take_old: bool
) -> Result<WriteOutcome, DispatchError>

Sets the storage entry by the given key to the specified value. If value is None then the storage entry is deleted.

fn caller(&self) -> &<Self::T as Config>::AccountId

Returns a reference to the account id of the caller.

fn is_contract(&self, address: &<Self::T as Config>::AccountId) -> bool

Check if a contract lives at the specified address.

fn caller_is_origin(&self) -> bool

Check if the caller of the current contract is the origin of the whole call stack.

This can be checked with is_contract(self.caller()) as well. However, this function does not require any storage lookup and therefore uses less weight.

fn address(&self) -> &<Self::T as Config>::AccountId

Returns a reference to the account id of the current contract.

fn balance(
&self
) -> <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance

Returns the balance of the current contract.

The value_transferred is already added.

fn value_transferred(
&self
) -> <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance

Returns the value transferred along with this call.

fn now(&self) -> &<<Self::T as Config>::Time as Time>::Moment

Returns a reference to the timestamp of the current block

fn minimum_balance(
&self
) -> <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance

Returns the minimum balance that is required for creating an account.

fn random(
&self,
subject: &[u8]
) -> (<Self::T as Config>::Hash, <Self::T as Config>::BlockNumber)

Returns a random number for the current block with the given subject.

fn deposit_event(
&mut self,
topics: Vec<<Self::T as Config>::Hash>,
data: Vec<u8>
)

Deposit an event with the given topics.

There should not be any duplicates in topics.

fn block_number(&self) -> <Self::T as Config>::BlockNumber

Returns the current block number.

fn max_value_size(&self) -> u32

Returns the maximum allowed size of a storage item.

fn get_weight_price(
&self,
weight: Weight
) -> <<Self::T as Config>::Currency as Currency<<Self::T as Config>::AccountId>>::Balance

Returns the price for the specified amount of weight.

fn schedule(&self) -> &Schedule<Self::T>

Get a reference to the schedule used by the current call.

fn gas_meter(&mut self) -> &mut GasMeter<Self::T>

Get a mutable reference to the nested gas meter.

fn append_debug_buffer(&mut self, msg: &str) -> bool

Append a string to the debug buffer.

It is added as-is without any additional new line.

This is a no-op if debug message recording is disabled which is always the case when the code is executing on-chain.

Returns true if debug message recording is enabled. Otherwise false is returned.

fn call_runtime(
&self,
call: <Self::T as Config>::Call
) -> DispatchResultWithPostInfo

Call some dispatchable and return the result.

fn ecdsa_recover(
&self,
signature: &[u8; 65],
message_hash: &[u8; 32]
) -> Result<[u8; 33], ()>

Recovers ECDSA compressed public key based on signature and message hash.