Trait pallet_contracts::chain_extension::RegisteredChainExtension

pub trait RegisteredChainExtension<C: Config>: ChainExtension<C> {
    const ID: u16;
}

A ChainExtension that can be composed with other extensions using a tuple.

An extension that implements this trait can be put in a tuple in order to have multiple extensions available. The tuple implementation routes requests based on the first two most significant bytes of the id passed to call.

If this extensions is to be used by multiple runtimes consider registering it to ensure that there are no collisions with other vendors.

Note

Currently, we support tuples of up to ten registered chain extensions. If more chain extensions are needed consider opening an issue.

Required Associated Constants

const ID: u16

The extensions globally unique identifier.

Object Safety

This trait is not object safe.