pallet_contracts

Module debug

Structs

ExecReturnValue

Output of a contract call or instantiation which ran to completion.

Enums

ExportedFunction

Describes the different functions that can be exported by an [Executable].

Traits

CallInterceptor

Provides an interface for intercepting contract calls.

CallSpan

Defines a span of execution for a contract call.

Debugger

Umbrella trait for all interfaces that serves for debugging.

Tracing

Defines methods to capture contract calls, enabling external observers to measure, trace, and react to contract interactions.

Type Aliases

ExecResult