Trait pallet_contracts::migration::MigrateSequence

pub trait MigrateSequence: Sealed {
    const VERSION_RANGE: (u16, u16);

    // Required methods
    fn new(version: StorageVersion) -> Cursor;
    fn steps(
        version: StorageVersion,
        cursor: &[u8],
        weight_left: &mut Weight
    ) -> StepResult;
    fn integrity_test(max_block_weight: Weight);

    // Provided methods
    fn pre_upgrade_step(
        _version: StorageVersion
    ) -> Result<Vec<u8>, TryRuntimeError> { ... }
    fn post_upgrade_step(
        _version: StorageVersion,
        _state: Vec<u8>
    ) -> Result<(), TryRuntimeError> { ... }
    fn is_upgrade_supported(
        in_storage: StorageVersion,
        target: StorageVersion
    ) -> bool { ... }
}

Defines a sequence of migrations.

The sequence must be defined by a tuple of migrations, each of which must implement the MigrationStep trait. Migrations must be ordered by their versions with no gaps.

Required Associated Constants

const VERSION_RANGE: (u16, u16)

Returns the range of versions that this migrations sequence can handle. Migrations must be ordered by their versions with no gaps.

The following code will fail to compile:

let _ = <(NoopMigration<1>, NoopMigration<3>)>::VERSION_RANGE;

The following code will compile:

let _ = <(NoopMigration<1>, NoopMigration<2>)>::VERSION_RANGE;

Required Methods

fn new(version: StorageVersion) -> Cursor

Returns the default cursor for the given version.

fn steps( version: StorageVersion, cursor: &[u8], weight_left: &mut Weight ) -> StepResult

Execute the migration step until the weight limit is reached.

fn integrity_test(max_block_weight: Weight)

Verify that the migration step fits into Cursor, and that max_step_weight is not greater than max_block_weight.

Provided Methods

fn pre_upgrade_step( _version: StorageVersion ) -> Result<Vec<u8>, TryRuntimeError>

fn post_upgrade_step( _version: StorageVersion, _state: Vec<u8> ) -> Result<(), TryRuntimeError>

fn is_upgrade_supported( in_storage: StorageVersion, target: StorageVersion ) -> bool

Returns whether migrating from in_storage to target is supported.

A migration is supported if VERSION_RANGE is (in_storage + 1, target).