pallet_contracts::migration

Module v15

Move contracts’ reserved balance from the deposit_account to be held in the contract’s account instead. Since Currency has been deprecated, we need the deposits to be handled by the frame_support::traits::fungible traits instead. For this transfer the balance from the deposit account to the contract’s account and hold it in there. Then the deposit account is not needed anymore and we can get rid of it.

Structs

Migration