Module pallet_contracts::pallet

Structs

GenesisConfig

Pallet

_GeneratedPrefixForStorageAccountCounter

_GeneratedPrefixForStorageCodeStorage

_GeneratedPrefixForStorageContractInfoOf

_GeneratedPrefixForStorageDeletionQueue

_GeneratedPrefixForStoragePristineCode

Enums

Call

Error

Event

Traits

Config

Type Definitions

AccountCounter

The subtrie counter.

CodeStorage

A mapping between an original code hash and instrumented wasm code, ready for execution.

ContractInfoOf

The code associated with a given account.

DeletionQueue

Evicted contracts that await child trie deletion.

Module

Type alias to Pallet, to be used by construct_runtime.

PristineCode

A mapping from an original code hash to the original code, untouched by instrumentation.