Module pallet_contracts::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

call

Makes a call to an account, optionally transferring some balance.

call_old_weight

Deprecated version if [Self::call] for use in an in-storage Call.

instantiate

Instantiates a contract from a previously deployed wasm binary.

instantiate_old_weight

Deprecated version if [Self::instantiate] for use in an in-storage Call.

instantiate_with_code

Instantiates a new contract from the supplied code optionally transferring some balance.

instantiate_with_code_old_weight

Deprecated version if [Self::instantiate_with_code] for use in an in-storage Call.

migrate

When a migration is in progress, this dispatchable can be used to run migration steps. Calls that contribute to advancing the migration have their fees waived, as it’s helpful for the chain. Note that while the migration is in progress, the pallet will also leverage the on_idle hooks to run migration steps.

remove_code

Remove the code stored under code_hash and refund the deposit to its owner.

set_code

Privileged function that changes the code of an existing contract.

upload_code

Upload new code without instantiating a contract from it.