Function pallet_contracts::pallet::dispatchables::call

pub fn call<T: Config>(
    dest: <<T as Config>::Lookup as StaticLookup>::Source,
    value: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance,
    gas_limit: Weight,
    storage_deposit_limit: Option<<<<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance as HasCompact>::Type>,
    data: Vec<u8>,
)

Makes a call to an account, optionally transferring some balance.

Parameters

dest: Address of the contract to call.

value: The balance to transfer from the origin to dest.

gas_limit: The gas limit enforced when executing the constructor.

storage_deposit_limit: The maximum amount of balance that can be charged from the caller to pay for the storage consumed.

data: The input data to pass to the contract.

If the account is a smart-contract account, the associated code will be executed and any value will be transferred.

If the account is a regular account, any value will be transferred.

If no account exists and the call value is not less than existential_deposit, a regular account will be created and any value will be transferred.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::call.