Module pallet_contracts::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

CodeInfoOf

A mapping from a contract’s code hash to its code info.

ContractInfoOf

The code associated with a given account.

DeletionQueue

Evicted contracts that await child trie deletion.

DeletionQueueCounter

A pair of monotonic counters used to track the latest contract marked for deletion and the latest deleted contract in queue.

MigrationInProgress

A migration can span across multiple blocks. This storage defines a cursor to track the progress of the migration, enabling us to resume from the last completed position.

Nonce

This is a monotonic counter incremented on contract instantiation.

PristineCode

A mapping from a contract’s code hash to its code.