Type Definition pallet_contracts::pallet::Module

pub type Module<T> = Pallet<T>;

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Implementations

impl<T: Config> Module<T> where
T::AccountId: UncheckedFrom<T::Hash> + AsRef<[u8]>,

pub fn bare_call(
origin: T::AccountId,
dest: T::AccountId,
value: BalanceOf<T>,
gas_limit: Weight,
input_data: Vec<u8>
) -> ContractExecResult

Perform a call to a specified contract.

This function is similar to Self::call, but doesn’t perform any address lookups and better suitable for calling directly from Rust.

It returns the execution result and the amount of used weight.

pub fn get_storage(address: T::AccountId, key: [u8; 32]) -> GetStorageResult

Query storage of a specified contract under a specified key.

pub fn rent_projection(
address: T::AccountId
) -> RentProjectionResult<T::BlockNumber>

Query how many blocks the contract stays alive given that the amount endowment and consumed storage does not change.

pub fn contract_address(
deploying_address: &T::AccountId,
code_hash: &CodeHash<T>,
salt: &[u8]
) -> T::AccountId

Determine the address of a contract,

This is the address generation function used by contract instantiation. Its result is only dependend on its inputs. It can therefore be used to reliably predict the address of a contract. This is akin to the formular of eth’s CREATE2 opcode. There is no CREATE equivalent because CREATE2 is strictly more powerful.

Formula: hash(deploying_address ++ code_hash ++ salt)

pub fn subsistence_threshold() -> BalanceOf<T>

Subsistence threshold is the extension of the minimum balance (aka existential deposit) by the tombstone deposit, required for leaving a tombstone.

Rent or any contract initiated balance transfer mechanism cannot make the balance lower than the subsistence threshold in order to guarantee that a tombstone is created.

The only way to completely kill a contract without a tombstone is calling seal_terminate.