Struct pallet_contracts::HostFnWeights

pub struct HostFnWeights<T: Config> {

    pub caller: Weight,
    pub is_contract: Weight,
    pub code_hash: Weight,
    pub own_code_hash: Weight,
    pub caller_is_origin: Weight,
    pub caller_is_root: Weight,
    pub address: Weight,
    pub gas_left: Weight,
    pub balance: Weight,
    pub value_transferred: Weight,
    pub minimum_balance: Weight,
    pub block_number: Weight,
    pub now: Weight,
    pub weight_to_fee: Weight,
    pub input: Weight,
    pub input_per_byte: Weight,
    pub return: Weight,
    pub return_per_byte: Weight,
    pub terminate: Weight,
    pub random: Weight,
    pub deposit_event: Weight,
    pub deposit_event_per_topic: Weight,
    pub deposit_event_per_byte: Weight,
    pub debug_message: Weight,
    pub debug_message_per_byte: Weight,
    pub set_storage: Weight,
    pub set_storage_per_new_byte: Weight,
    pub set_storage_per_old_byte: Weight,
    pub set_code_hash: Weight,
    pub clear_storage: Weight,
    pub clear_storage_per_byte: Weight,
    pub contains_storage: Weight,
    pub contains_storage_per_byte: Weight,
    pub get_storage: Weight,
    pub get_storage_per_byte: Weight,
    pub take_storage: Weight,
    pub take_storage_per_byte: Weight,
    pub transfer: Weight,
    pub call: Weight,
    pub delegate_call: Weight,
    pub call_transfer_surcharge: Weight,
    pub call_per_cloned_byte: Weight,
    pub instantiate: Weight,
    pub instantiate_transfer_surcharge: Weight,
    pub instantiate_per_input_byte: Weight,
    pub instantiate_per_salt_byte: Weight,
    pub hash_sha2_256: Weight,
    pub hash_sha2_256_per_byte: Weight,
    pub hash_keccak_256: Weight,
    pub hash_keccak_256_per_byte: Weight,
    pub hash_blake2_256: Weight,
    pub hash_blake2_256_per_byte: Weight,
    pub hash_blake2_128: Weight,
    pub hash_blake2_128_per_byte: Weight,
    pub ecdsa_recover: Weight,
    pub ecdsa_to_eth_address: Weight,
    pub sr25519_verify: Weight,
    pub sr25519_verify_per_byte: Weight,
    pub reentrance_count: Weight,
    pub account_reentrance_count: Weight,
    pub instantiation_nonce: Weight,
    pub add_delegate_dependency: Weight,
    pub remove_delegate_dependency: Weight,
    pub _phantom: PhantomData<T>,

}

Describes the weight for each imported function that a contract is allowed to call.

Fields
caller: Weight

Weight of calling seal_caller.

is_contract: Weight

Weight of calling seal_is_contract.

code_hash: Weight

Weight of calling seal_code_hash.

own_code_hash: Weight

Weight of calling seal_own_code_hash.

caller_is_origin: Weight

Weight of calling seal_caller_is_origin.

caller_is_root: Weight

Weight of calling seal_caller_is_root.

address: Weight

Weight of calling seal_address.

gas_left: Weight

Weight of calling seal_gas_left.

balance: Weight

Weight of calling seal_balance.

value_transferred: Weight

Weight of calling seal_value_transferred.

minimum_balance: Weight

Weight of calling seal_minimum_balance.

block_number: Weight

Weight of calling seal_block_number.

now: Weight

Weight of calling seal_now.

weight_to_fee: Weight

Weight of calling seal_weight_to_fee.

input: Weight

Weight of calling seal_input.

input_per_byte: Weight

Weight per input byte copied to contract memory by seal_input.

return: Weight

Weight of calling seal_return.

return_per_byte: Weight

Weight per byte returned through seal_return.

terminate: Weight

Weight of calling seal_terminate.

random: Weight

Weight of calling seal_random.

deposit_event: Weight

Weight of calling seal_reposit_event.

deposit_event_per_topic: Weight

Weight per topic supplied to seal_deposit_event.

deposit_event_per_byte: Weight

Weight per byte of an event deposited through seal_deposit_event.

debug_message: Weight

Weight of calling seal_debug_message.

debug_message_per_byte: Weight

Weight of calling seal_debug_message per byte of the message.

set_storage: Weight

Weight of calling seal_set_storage.

set_storage_per_new_byte: Weight

Weight per written byten of an item stored with seal_set_storage.

set_storage_per_old_byte: Weight

Weight per overwritten byte of an item stored with seal_set_storage.

set_code_hash: Weight

Weight of calling seal_set_code_hash.

clear_storage: Weight

Weight of calling seal_clear_storage.

clear_storage_per_byte: Weight

Weight of calling seal_clear_storage per byte of the stored item.

contains_storage: Weight

Weight of calling seal_contains_storage.

contains_storage_per_byte: Weight

Weight of calling seal_contains_storage per byte of the stored item.

get_storage: Weight

Weight of calling seal_get_storage.

get_storage_per_byte: Weight

Weight per byte of an item received via seal_get_storage.

take_storage: Weight

Weight of calling seal_take_storage.

take_storage_per_byte: Weight

Weight per byte of an item received via seal_take_storage.

transfer: Weight

Weight of calling seal_transfer.

call: Weight

Weight of calling seal_call.

delegate_call: Weight

Weight of calling seal_delegate_call.

call_transfer_surcharge: Weight

Weight surcharge that is claimed if seal_call does a balance transfer.

call_per_cloned_byte: Weight

Weight per byte that is cloned by supplying the CLONE_INPUT flag.

instantiate: Weight

Weight of calling seal_instantiate.

instantiate_transfer_surcharge: Weight

Weight surcharge that is claimed if seal_instantiate does a balance transfer.

instantiate_per_input_byte: Weight

Weight per input byte supplied to seal_instantiate.

instantiate_per_salt_byte: Weight

Weight per salt byte supplied to seal_instantiate.

hash_sha2_256: Weight

Weight of calling seal_hash_sha_256.

hash_sha2_256_per_byte: Weight

Weight per byte hashed by seal_hash_sha_256.

hash_keccak_256: Weight

Weight of calling seal_hash_keccak_256.

hash_keccak_256_per_byte: Weight

Weight per byte hashed by seal_hash_keccak_256.

hash_blake2_256: Weight

Weight of calling seal_hash_blake2_256.

hash_blake2_256_per_byte: Weight

Weight per byte hashed by seal_hash_blake2_256.

hash_blake2_128: Weight

Weight of calling seal_hash_blake2_128.

hash_blake2_128_per_byte: Weight

Weight per byte hashed by seal_hash_blake2_128.

ecdsa_recover: Weight

Weight of calling seal_ecdsa_recover.

ecdsa_to_eth_address: Weight

Weight of calling seal_ecdsa_to_eth_address.

sr25519_verify: Weight

Weight of calling sr25519_verify.

sr25519_verify_per_byte: Weight

Weight per byte of calling sr25519_verify.

reentrance_count: Weight

Weight of calling reentrance_count.

account_reentrance_count: Weight

Weight of calling account_reentrance_count.

instantiation_nonce: Weight

Weight of calling instantiation_nonce.

add_delegate_dependency: Weight

Weight of calling add_delegate_dependency.

remove_delegate_dependency: Weight

Weight of calling remove_delegate_dependency.

_phantom: PhantomData<T>

The type parameter is used in the default implementation.