Struct pallet_contracts::InstructionWeights

pub struct InstructionWeights<T: Config> {
    pub base: u32,
    pub _phantom: PhantomData<T>,
}

Gas metering of Wasm executed instructions is being done on the engine side. This struct holds a reference value used to gas units scaling between host and engine.

Fields
base: u32

Base instruction ref_time Weight. Should match to wasmi’s 1 fuel (see https://github.com/paritytech/wasmi/issues/701).

_phantom: PhantomData<T>

The type parameter is used in the default implementation.