Module pallet_contracts::test_utils

Shared utilities for testing contracts. This is not part of the tests module because it is made public for other crates to use.

Modules

builder

Structs

AccountId32

An opaque 32-byte cryptographic identifier.

Constants

ALICE

BOB

CHARLIE

DJANGO

GAS_LIMIT