Crate pallet_contracts

Contract Pallet

The Contract module provides functionality for the runtime to deploy and execute WebAssembly smart-contracts.

Config

Call

Overview

This module extends accounts based on the Currency trait to have smart-contract functionality. It can be used with other modules that implement accounts based on Currency. These “smart-contract accounts” have the ability to instantiate smart-contracts and make calls to other contract and non-contract accounts.

The smart-contract code is stored once in a code cache, and later retrievable via its hash. This means that multiple smart-contracts can be instantiated from the same hash, without replicating the code each time.

When a smart-contract is called, its associated code is retrieved via the code hash and gets executed. This call can alter the storage entries of the smart-contract account, instantiate new smart-contracts, or call other smart-contracts.

Finally, when an account is reaped, its associated code and storage of the smart-contract account will also be deleted.

Gas

Senders must specify a gas limit with every call, as all instructions invoked by the smart-contract require gas. Unused gas is refunded after the call, regardless of the execution outcome.

If the gas limit is reached, then all calls and state changes (including balance transfers) are only reverted at the current call’s contract level. For example, if contract A calls B and B runs out of gas mid-call, then all of B’s calls are reverted. Assuming correct error handling by contract A, A’s other calls and state changes still persist.

Notable Scenarios

Contract call failures are not always cascading. When failures occur in a sub-call, they do not “bubble up”, and the call will only revert at the specific contract level. For example, if contract A calls contract B, and B fails, A can decide how to handle that failure, either proceeding or reverting A’s changes.

Interface

Dispatchable functions

Pallet::instantiate_with_code - Deploys a new contract from the supplied wasm binary, optionally transferring some balance. This instantiates a new smart contract account with the supplied code and calls its constructor to initialize the contract.

Pallet::instantiate - The same as instantiate_with_code but instead of uploading new code an existing code_hash is supplied.

Pallet::call - Makes a call to an account, optionally transferring some balance.

Usage

The Contract module is a work in progress. The following examples show how this Contract module can be used to instantiate and call contracts.

ink is an eDSL that enables writing WebAssembly based smart contracts in the Rust programming language. This is a work in progress.

Re-exports

pub use crate::pallet::*;

Modules

chain_extension

A mechanism for runtime authors to augment the functionality of contracts.

migration

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_contracts

Structs

DefaultAddressGenerator

Default address generator.

Frame

Represents one entry in the call stack.

HostFnWeights

Describes the weight for each imported function that a contract is allowed to call.

InstructionWeights

Describes the weight for all categories of supported wasm instructions.

Limits

Describes the upper limits on various metrics.

Schedule

Definition of the cost schedule and other parameterizations for the wasm vm.

Traits

AddressGenerator

Provides the contract address generation method.