Module pallet_contracts::api_doc

Documentation of the API (host functions) available to contracts.

The Current trait might be the most useful doc to look at. The versioned traits only exist for reference: If trying to find out if a specific version of pallet-contracts contains a certain function.

Note

This module is not meant to be used by any code. Rather, it is meant to be consumed by humans through rustdoc.

Traits

Current

Contains only the latest version of each function.

Version0

All functions available in the seal0 module

Version1

All functions available in the seal1 module

Version2

All functions available in the seal2 module