Trait pallet_contracts::chain_extension::Ext

pub trait Ext {
    type T: Config;
    pub fn get_storage(&self, key: &[u8; 32]) -> Option<Vec<u8>>;

    pub fn set_storage(
        &mut self, 
        key: [u8; 32], 
        value: Option<Vec<u8>>
    ) -> DispatchResult;

    pub fn instantiate(
        &mut self, 
        code: CodeHash<Self::T>, 
        value: BalanceOf<Self::T>, 
        gas_meter: &mut GasMeter<Self::T>, 
        input_data: Vec<u8>, 
        salt: &[u8]
    ) -> Result<(<Self::T as Config>::AccountId, ExecReturnValue), ExecError>;

    pub fn transfer(
        &mut self, 
        to: &<Self::T as Config>::AccountId, 
        value: BalanceOf<Self::T>
    ) -> DispatchResult;

    pub fn terminate(
        &mut self, 
        beneficiary: &<Self::T as Config>::AccountId
    ) -> DispatchResult;

    pub fn call(
        &mut self, 
        to: &<Self::T as Config>::AccountId, 
        value: BalanceOf<Self::T>, 
        gas_meter: &mut GasMeter<Self::T>, 
        input_data: Vec<u8>
    ) -> ExecResult;

    pub fn restore_to(
        &mut self, 
        dest: <Self::T as Config>::AccountId, 
        code_hash: CodeHash<Self::T>, 
        rent_allowance: BalanceOf<Self::T>, 
        delta: Vec<[u8; 32]>
    ) -> DispatchResult;

    pub fn caller(&self) -> &<Self::T as Config>::AccountId;

    pub fn address(&self) -> &<Self::T as Config>::AccountId;

    pub fn balance(&self) -> BalanceOf<Self::T>;

    pub fn value_transferred(&self) -> BalanceOf<Self::T>;

    pub fn now(&self) -> &<<Self::T as Config>::Time as Time>::Moment;

    pub fn minimum_balance(&self) -> BalanceOf<Self::T>;

    pub fn tombstone_deposit(&self) -> BalanceOf<Self::T>;

    pub fn random(&self, subject: &[u8]) -> <Self::T as Config>::Hash;

    pub fn deposit_event(
        &mut self, 
        topics: Vec<<Self::T as Config>::Hash>, 
        data: Vec<u8>
    );

    pub fn set_rent_allowance(&mut self, rent_allowance: BalanceOf<Self::T>);

    pub fn rent_allowance(&self) -> BalanceOf<Self::T>;

    pub fn block_number(&self) -> <Self::T as Config>::BlockNumber;

    pub fn max_value_size(&self) -> u32;

    pub fn get_weight_price(&self, weight: Weight) -> BalanceOf<Self::T>;

    pub fn schedule(&self) -> &Schedule<Self::T>;
}

An interface that provides access to the external environment in which the smart-contract is executed.

This interface is specialized to an account of the executing code, so all operations are implicitly performed on that account.

Associated Types

type T: Config

Loading content...
Required methods

pub fn get_storage(&self, key: &[u8; 32]) -> Option<Vec<u8>>

Returns the storage entry of the executing account by the given key.

Returns None if the key wasn't previously set by set_storage or was deleted.

pub fn set_storage(
&mut self,
key: [u8; 32],
value: Option<Vec<u8>>
) -> DispatchResult

Sets the storage entry by the given key to the specified value. If value is None then the storage entry is deleted.

pub fn instantiate(
&mut self,
code: CodeHash<Self::T>,
value: BalanceOf<Self::T>,
gas_meter: &mut GasMeter<Self::T>,
input_data: Vec<u8>,
salt: &[u8]
) -> Result<(<Self::T as Config>::AccountId, ExecReturnValue), ExecError>

Instantiate a contract from the given code.

The newly created account will be associated with code. value specifies the amount of value transferred from this to the newly created account (also known as endowment).

pub fn transfer(
&mut self,
to: &<Self::T as Config>::AccountId,
value: BalanceOf<Self::T>
) -> DispatchResult

Transfer some amount of funds into the specified account.

pub fn terminate(
&mut self,
beneficiary: &<Self::T as Config>::AccountId
) -> DispatchResult

Transfer all funds to beneficiary and delete the contract.

Since this function removes the self contract eagerly, if succeeded, no further actions should be performed on this Ext instance.

This function will fail if the same contract is present on the contract call stack.

pub fn call(
&mut self,
to: &<Self::T as Config>::AccountId,
value: BalanceOf<Self::T>,
gas_meter: &mut GasMeter<Self::T>,
input_data: Vec<u8>
) -> ExecResult

Call (possibly transferring some amount of funds) into the specified account.

pub fn restore_to(
&mut self,
dest: <Self::T as Config>::AccountId,
code_hash: CodeHash<Self::T>,
rent_allowance: BalanceOf<Self::T>,
delta: Vec<[u8; 32]>
) -> DispatchResult

Restores the given destination contract sacrificing the current one.

Since this function removes the self contract eagerly, if succeeded, no further actions should be performed on this Ext instance.

This function will fail if the same contract is present on the contract call stack.

pub fn caller(&self) -> &<Self::T as Config>::AccountId

Returns a reference to the account id of the caller.

pub fn address(&self) -> &<Self::T as Config>::AccountId

Returns a reference to the account id of the current contract.

pub fn balance(&self) -> BalanceOf<Self::T>

Returns the balance of the current contract.

The value_transferred is already added.

pub fn value_transferred(&self) -> BalanceOf<Self::T>

Returns the value transferred along with this call or as endowment.

pub fn now(&self) -> &<<Self::T as Config>::Time as Time>::Moment

Returns a reference to the timestamp of the current block

pub fn minimum_balance(&self) -> BalanceOf<Self::T>

Returns the minimum balance that is required for creating an account.

pub fn tombstone_deposit(&self) -> BalanceOf<Self::T>

Returns the deposit required to create a tombstone upon contract eviction.

pub fn random(&self, subject: &[u8]) -> <Self::T as Config>::Hash

Returns a random number for the current block with the given subject.

pub fn deposit_event(
&mut self,
topics: Vec<<Self::T as Config>::Hash>,
data: Vec<u8>
)

Deposit an event with the given topics.

There should not be any duplicates in topics.

pub fn set_rent_allowance(&mut self, rent_allowance: BalanceOf<Self::T>)

Set rent allowance of the contract

pub fn rent_allowance(&self) -> BalanceOf<Self::T>

Rent allowance of the contract

pub fn block_number(&self) -> <Self::T as Config>::BlockNumber

Returns the current block number.

pub fn max_value_size(&self) -> u32

Returns the maximum allowed size of a storage item.

pub fn get_weight_price(&self, weight: Weight) -> BalanceOf<Self::T>

Returns the price for the specified amount of weight.

pub fn schedule(&self) -> &Schedule<Self::T>

Get a reference to the schedule used by the current call.

Loading content...