Trait pallet_contracts::migration::MigrationStep

pub trait MigrationStep: Codec + MaxEncodedLen + Default {
    const VERSION: u16;

    // Required methods
    fn max_step_weight() -> Weight;
    fn step(&mut self) -> (IsFinished, Weight);

    // Provided method
    fn integrity_test(max_block_weight: Weight) { ... }
}

A trait that allows to migrate storage from one version to another.

The migration is done in steps. The migration is finished when step() returns IsFinished::Yes.

Required Associated Constants

const VERSION: u16

Returns the version of the migration.

Required Methods

fn max_step_weight() -> Weight

Returns the maximum weight that can be consumed in a single step.

fn step(&mut self) -> (IsFinished, Weight)

Process one step of the migration.

Returns whether the migration is finished and the weight consumed.

Provided Methods

fn integrity_test(max_block_weight: Weight)

Verify that the migration step fits into Cursor, and that max_step_weight is not greater than max_block_weight.

Object Safety

This trait is not object safe.