Trait pallet_contracts::pallet::Config

pub trait Config: Config {
    type Time: Time;
    type Randomness: Randomness<Self::Hash, Self::BlockNumber>;
    type Currency: Currency<Self::AccountId>;
    type Event: From<Event<Self>> + IsType<Self::Event>;
    type RentPayment: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type WeightPrice: Convert<Weight, <Self::Currency as Currency<Self::AccountId>>::Balance>;
    type WeightInfo: WeightInfo;
    type ChainExtension: ChainExtension<Self>;
    type Schedule: Get<Schedule<Self>>;
    type SignedClaimHandicap: Get<Self::BlockNumber>;
    type TombstoneDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type DepositPerContract: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type DepositPerStorageByte: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type DepositPerStorageItem: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type RentFraction: Get<Perbill>;
    type SurchargeReward: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type CallStack: Array<Item = Frame<Self>>;
    type DeletionQueueDepth: Get<u32>;
    type DeletionWeightLimit: Get<Weight>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Time: Time

The time implementation used to supply timestamps to conntracts through seal_now.

type Randomness: Randomness<Self::Hash, Self::BlockNumber>

The generator used to supply randomness to contracts through seal_random.

type Currency: Currency<Self::AccountId>

The currency in which fees are paid and contract balances are held.

type Event: From<Event<Self>> + IsType<Self::Event>

The overarching event type.

type RentPayment: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

Handler for rent payments.

type WeightPrice: Convert<Weight, <Self::Currency as Currency<Self::AccountId>>::Balance>

Used to answer contracts’ queries regarding the current weight price. This is not used to calculate the actual fee and is only for informational purposes.

type WeightInfo: WeightInfo

Describes the weights of the dispatchables of this module and is also used to construct a default cost schedule.

type ChainExtension: ChainExtension<Self>

Type that allows the runtime authors to add new host functions for a contract to call.

type Schedule: Get<Schedule<Self>>

Cost schedule and limits.

type SignedClaimHandicap: Get<Self::BlockNumber>

Number of block delay an extrinsic claim surcharge has.

When claim surcharge is called by an extrinsic the rent is checked for current_block - delay

type TombstoneDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The minimum amount required to generate a tombstone.

type DepositPerContract: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The balance every contract needs to deposit to stay alive indefinitely.

This is different from the Self::TombstoneDeposit because this only needs to be deposited while the contract is alive. Costs for additional storage are added to this base cost.

This is a simple way to ensure that contracts with empty storage eventually get deleted by making them pay rent. This creates an incentive to remove them early in order to save rent.

type DepositPerStorageByte: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The balance a contract needs to deposit per storage byte to stay alive indefinitely.

Let’s suppose the deposit is 1,000 BU (balance units)/byte and the rent is 1 BU/byte/day, then a contract with 1,000,000 BU that uses 1,000 bytes of storage would pay no rent. But if the balance reduced to 500,000 BU and the storage stayed the same at 1,000, then it would pay 500 BU/day.

type DepositPerStorageItem: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The balance a contract needs to deposit per storage item to stay alive indefinitely.

It works the same as Self::DepositPerStorageByte but for storage items.

type RentFraction: Get<Perbill>

The fraction of the deposit that should be used as rent per block.

When a contract hasn’t enough balance deposited to stay alive indefinitely it needs to pay per block for the storage it consumes that is not covered by the deposit. This determines how high this rent payment is per block as a fraction of the deposit.

type SurchargeReward: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

Reward that is received by the party whose touch has led to removal of a contract.

type CallStack: Array<Item = Frame<Self>>

The type of the call stack determines the maximum nesting depth of contract calls.

The allowed depth is CallStack::size() + 1. Therefore a size of 0 means that a contract cannot use call or instantiate. In other words only the origin called “root contract” is allowed to execute then.

type DeletionQueueDepth: Get<u32>

The maximum number of tries that can be queued for deletion.

type DeletionWeightLimit: Get<Weight>

The maximum amount of weight that can be consumed per block for lazy trie removal.

Loading content...