Trait pallet_contracts::pallet::Config

pub trait Config: Config {

    type Time: Time;
    type Randomness: Randomness<Self::Hash>;
    type Currency: Currency<Self::AccountId>;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type RentPayment: OnUnbalanced<NegativeImbalanceOf<Self>>;
    type SignedClaimHandicap: Get<Self::BlockNumber>;
    type TombstoneDeposit: Get<BalanceOf<Self>>;
    type DepositPerContract: Get<BalanceOf<Self>>;
    type DepositPerStorageByte: Get<BalanceOf<Self>>;
    type DepositPerStorageItem: Get<BalanceOf<Self>>;
    type RentFraction: Get<Perbill>;
    type SurchargeReward: Get<BalanceOf<Self>>;
    type MaxDepth: Get<u32>;
    type MaxValueSize: Get<u32>;
    type WeightPrice: Convert<Weight, BalanceOf<Self>>;
    type WeightInfo: WeightInfo;
    type ChainExtension: ChainExtension<Self>;
    type DeletionQueueDepth: Get<u32>;
    type DeletionWeightLimit: Get<Weight>;
    type MaxCodeSize: Get<u32>;

}

Associated Types

type Time: Time

The time implementation used to supply timestamps to conntracts through seal_now.

type Randomness: Randomness<Self::Hash>

The generator used to supply randomness to contracts through seal_random.

type Currency: Currency<Self::AccountId>

The currency in which fees are paid and contract balances are held.

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type RentPayment: OnUnbalanced<NegativeImbalanceOf<Self>>

Handler for rent payments.

type SignedClaimHandicap: Get<Self::BlockNumber>

Number of block delay an extrinsic claim surcharge has.

When claim surcharge is called by an extrinsic the rent is checked for current_block - delay

type TombstoneDeposit: Get<BalanceOf<Self>>

The minimum amount required to generate a tombstone.

type DepositPerContract: Get<BalanceOf<Self>>

The balance every contract needs to deposit to stay alive indefinitely.

This is different from the Self::TombstoneDeposit because this only needs to be deposited while the contract is alive. Costs for additional storage are added to this base cost.

This is a simple way to ensure that contracts with empty storage eventually get deleted by making them pay rent. This creates an incentive to remove them early in order to save rent.

type DepositPerStorageByte: Get<BalanceOf<Self>>

The balance a contract needs to deposit per storage byte to stay alive indefinitely.

Let’s suppose the deposit is 1,000 BU (balance units)/byte and the rent is 1 BU/byte/day, then a contract with 1,000,000 BU that uses 1,000 bytes of storage would pay no rent. But if the balance reduced to 500,000 BU and the storage stayed the same at 1,000, then it would pay 500 BU/day.

type DepositPerStorageItem: Get<BalanceOf<Self>>

The balance a contract needs to deposit per storage item to stay alive indefinitely.

It works the same as Self::DepositPerStorageByte but for storage items.

type RentFraction: Get<Perbill>

The fraction of the deposit that should be used as rent per block.

When a contract hasn’t enough balance deposited to stay alive indefinitely it needs to pay per block for the storage it consumes that is not covered by the deposit. This determines how high this rent payment is per block as a fraction of the deposit.

type SurchargeReward: Get<BalanceOf<Self>>

Reward that is received by the party whose touch has led to removal of a contract.

type MaxDepth: Get<u32>

The maximum nesting level of a call/instantiate stack.

type MaxValueSize: Get<u32>

The maximum size of a storage value and event payload in bytes.

type WeightPrice: Convert<Weight, BalanceOf<Self>>

Used to answer contracts’s queries regarding the current weight price. This is not used to calculate the actual fee and is only for informational purposes.

type WeightInfo: WeightInfo

Describes the weights of the dispatchables of this module and is also used to construct a default cost schedule.

type ChainExtension: ChainExtension<Self>

Type that allows the runtime authors to add new host functions for a contract to call.

type DeletionQueueDepth: Get<u32>

The maximum number of tries that can be queued for deletion.

type DeletionWeightLimit: Get<Weight>

The maximum amount of weight that can be consumed per block for lazy trie removal.

type MaxCodeSize: Get<u32>

The maximum length of a contract code in bytes. This limit applies to the instrumented version of the code. Therefore instantiate_with_code can fail even when supplying a wasm binary below this maximum size.