Trait pallet_contracts::pallet::Config

pub trait Config: Config {

    type Time: Time;
    type Randomness: Randomness<Self::Hash, Self::BlockNumber>;
    type Currency: Currency<Self::AccountId>;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Call: Dispatchable<Origin = Self::Origin, PostInfo = PostDispatchInfo> + GetDispatchInfo + Decode + IsType<<Self as Config>::Call>;
    type CallFilter: Contains<<Self as Config>::Call>;
    type WeightPrice: Convert<Weight, <<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type WeightInfo: WeightInfo;
    type ChainExtension: ChainExtension<Self>;
    type Schedule: Get<Schedule<Self>>;
    type ContractDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type CallStack: Array<Item = Frame<Self>>;
    type DeletionQueueDepth: Get<u32>;
    type DeletionWeightLimit: Get<Weight>;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Time: Time

The time implementation used to supply timestamps to conntracts through seal_now.

type Randomness: Randomness<Self::Hash, Self::BlockNumber>

The generator used to supply randomness to contracts through seal_random.

type Currency: Currency<Self::AccountId>

The currency in which fees are paid and contract balances are held.

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type Call: Dispatchable<Origin = Self::Origin, PostInfo = PostDispatchInfo> + GetDispatchInfo + Decode + IsType<<Self as Config>::Call>

The overarching call type.

type CallFilter: Contains<<Self as Config>::Call>

Filter that is applied to calls dispatched by contracts.

Use this filter to control which dispatchables are callable by contracts. This is applied in addition to frame_system::Config::BaseCallFilter. It is recommended to treat this as a whitelist.

Subsistence Threshold

The runtime must make sure that any allowed dispatchable makes sure that the total_balance of the contract stays above Pallet::subsistence_threshold(). Otherwise users could clutter the storage with contracts.

Stability

The runtime must make sure that all dispatchables that are callable by contracts remain stable. In addition Self::Call itself must remain stable. This means that no existing variants are allowed to switch their positions.

Note

Note that dispatchables that are called via contracts do not spawn their own wasm instance for each call (as opposed to when called via a transaction). Therefore please make sure to be restrictive about which dispatchables are allowed in order to not introduce a new DoS vector like memory allocation patterns that can be exploited to drive the runtime into a panic.

type WeightPrice: Convert<Weight, <<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Used to answer contracts’ queries regarding the current weight price. This is not used to calculate the actual fee and is only for informational purposes.

type WeightInfo: WeightInfo

Describes the weights of the dispatchables of this module and is also used to construct a default cost schedule.

type ChainExtension: ChainExtension<Self>

Type that allows the runtime authors to add new host functions for a contract to call.

type Schedule: Get<Schedule<Self>>

Cost schedule and limits.

type ContractDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The deposit that must be placed into the contract’s account to instantiate it. This is in addition to the Currency::minimum_balance. The minimum balance for a contract’s account can be queried using Pallet::subsistence_threshold.

type CallStack: Array<Item = Frame<Self>>

The type of the call stack determines the maximum nesting depth of contract calls.

The allowed depth is CallStack::size() + 1. Therefore a size of 0 means that a contract cannot use call or instantiate. In other words only the origin called “root contract” is allowed to execute then.

type DeletionQueueDepth: Get<u32>

The maximum number of tries that can be queued for deletion.

type DeletionWeightLimit: Get<Weight>

The maximum amount of weight that can be consumed per block for lazy trie removal.