Trait pallet_contracts::AddressGenerator

pub trait AddressGenerator<T: Config> {
    fn generate_address(
        deploying_address: &T::AccountId,
        code_hash: &<T as Config>::Hash,
        salt: &[u8]
    ) -> T::AccountId;
}

Provides the contract address generation method.

See DefaultAddressGenerator for the default implementation.

Required Methods

fn generate_address(
deploying_address: &T::AccountId,
code_hash: &<T as Config>::Hash,
salt: &[u8]
) -> T::AccountId

Generate the address of a contract based on the given instantiate parameters.

Note for implementors

Make sure that there are no collisions, different inputs never lead to the same output.

Make sure that the same inputs lead to the same output.

Changing the implementation through a runtime upgrade without a proper storage migration would lead to catastrophic misbehavior.